some tidying of the test folder
